connectors/sentinelone2tone/pyproject.toml
[project]
name = "sentinelone2tone"
version = "0.9.7"
description = "SentinelOne Singularity -> Tenable One"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "arrow>=1.3.0",
    "restfly>=1.5.0",
    "rich>=13.9.3",
    "tomlkit>=0.13.2",
    "typer>=0.12.5",
    "pytenable>=1.7.1",
    "pydantic-extra-types>=2.9.0",
    "sqlalchemy>=2.0.36",
    "tenint>=0.1.3",
]

[project.urls]
support = "https://community.tenable.com"

[[project.authors]]
name = "Tenable, Inc."
email = "ecosystem-support@tenable.com"

[project.optional-dependencies]
testing = [
    "pytest>=8.3.3",
    "pytest-cov>=6.0.0",
    "responses>=0.25.3",
]


[tool.tenint.connector]
title = "SentinelOne to Tenable One"
tags = ["sentinelone", "tvm"]

[tool.tenint.connector.timeout]
default = 14400

[tool.ruff]
line-length = 88
indent-width = 4
exclude = [
  ".nova",
  ".github",
  ".git",
  ".pytest_cache",
  "__pycache__"
]

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "B"]
fixable = [ "ALL" ]
unfixable = [ "B" ]

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
line-ending = "lf"
docstring-code-format = false
docstring-code-line-length = "dynamic"

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E402", "F401"]
"**/{tests,docs,tools}/*" = ["E402"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
addopts = "--cov=sentinelone --cov-report=term-missing"
filterwarnings = ["ignore::DeprecationWarning"]

[tool.bandit]
exclude_dirs = ["tests", ".venv"]
